+++ src/components/Card/src/Card.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts" name="Card">
   withDefaults(defineProps<{
-    title?: string,
-    height?: string,
+    title?: string
+    height?: string
     padding?: boolean
   }>(), {
     height: 'auto',
@@ -11,11 +11,13 @@
 
 <template>
   <div bg="page" rounded :style="{ height, padding: padding ? '16px 24px' : '0' }">
-    <div v-if="title" center justify-between>
-      <h5 m="0" text="regular xl">{{ title }}</h5>
-      <slot name="right"></slot>
+    <div v-if="title" flex="center" justify="between">
+      <h5 m="0" text="regular xl">
+        {{ title }}
+      </h5>
+      <slot name="right" />
     </div>
-    <slot></slot>
+    <slot />
   </div>
 </template>