+++ src/components/Splitter/src/Trigger.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts" name="DraggleBar">
   const props = withDefaults(defineProps<{
-    direction?: 'row' | 'column',
+    direction?: 'row' | 'column'
     draggable?: boolean
     size?: number
   }>(), {
@@ -11,27 +11,29 @@
 
   function computedStyle(direction: string, isBtn = false) {
     const length = isBtn ? '36px' : '100%'
-    return direction === 'row' ? {
-      width: length,
-      height: `${props.size}px`,
-      cursor: isBtn ? 'ns-resize' : ''
-    } : {
-      width: `${props.size}px`,
-      height: length,
-      cursor: isBtn ? 'ew-resize' : ''
-    }
+    return direction === 'row'
+      ? {
+        width: length,
+        height: `${props.size}px`,
+        cursor: isBtn ? 'ns-resize' : ''
+      }
+      : {
+        width: `${props.size}px`,
+        height: length,
+        cursor: isBtn ? 'ew-resize' : ''
+      }
   }
 </script>
 
 <template>
-  <div :style="computedStyle(direction)" bg-root center>
+  <div :style="computedStyle(direction)" bg-root flex="center">
     <div
       v-if="draggable"
       :style="computedStyle(direction, true)"
       bg="placeholder hover:secondary active:regular"
       transition="base colors"
       rounded-full
-      center
-    ></div>
+      flex-center
+    />
   </div>
 </template>