+++ src/layouts/admin/sider/components/LogoView.vue
@@ -5,19 +5,16 @@
   const settingStore = useSettingStore()
 
   const router = useRouter()
-
 </script>
 
 <template>
   <div
     h="header"
-    flex
-    justify="center"
-    items="center"
+    flex="center"
     cursor="pointer"
     @click="router.push('/home')"
   >
-    <i-app-logo text="3xl" />
+    <img src="@/assets/icons/logo.svg" width="28">
     <h1
       v-show="!settingStore.menuCollapsed"
       text="xl main hover:regular"