+++ src/locales/lang/zh-cn.ts
@@ -36,7 +36,7 @@ export default {
       user: '用户管理',
       userDetail: '用户详情',
       department: '部门管理',
-      post: '岗位管理',
+      post: '岗位管理'
     },
     component: {
       root: '组件',
@@ -44,7 +44,8 @@ export default {
       table: '表格',
       splitPane: '分栏器',
       icon: '图标',
-      editor: '编辑器'
+      editor: '编辑器',
+      verificationCode: '验证码'
     },
     feat: {
       root: '功能',