+++ src/utils/http/helper.ts
@@ -1,18 +1,15 @@
 import type { AxiosResponse } from 'axios'
+import { VITE_BASE_API, VITE_USE_PROXY, VITE_USE_MOCK } from 'vite-env'
 import type { RequestConfig, Result } from '#/http'
 import { ErrorCodeEnum, HttpMethodEnum, CodeEnum } from '@/enums/httpEnum'
 import { alertErrMsg } from '@/utils/message'
 import { isDevMode } from '@/utils/env'
-import { envParse } from '../../../build/utils'
 
 // 生成token
 export const generateBaseToken = (token: Nullable<string>) => `bearer ${token}`
 
 // 获取默认请求前缀地址
 export const getDefaultBaseURL = () => {
-  const { VITE_BASE_API, VITE_USE_PROXY, VITE_USE_MOCK } = envParse(
-    import.meta.env
-  )
   // 使用proxy或者mock时，无需设置默认请求前缀
   if (isDevMode() && (VITE_USE_PROXY || VITE_USE_MOCK)) {
     return ''
@@ -42,7 +39,8 @@ export const transformResponse = (
   const { code, data, message: msg } = res.data
   if (code === CodeEnum.SUCCESS) {
     return data
-  } else {
+  }
+  else {
     alertErrMsg(`${ErrorCodeEnum.B}${code}`, msg)
     throw new Error(code.toString())
   }