+++ src/views/login/components/LoginForm.vue
@@ -1,15 +1,15 @@
 <script setup lang="ts" name="LoginForm">
   import { reactive } from 'vue'
-  import config from '@/config'
-  import { LoginParams } from '@/api/_auth/model'
   import { useMessage } from '@h/web/useMessage'
   import { useLoginByPassword } from '@h/logic/useLogin'
-  import { FormInstance, FormRules } from 'element-plus'
-  import { useRouter } from 'vue-router'
+  import type { FormInstance, FormRules } from 'element-plus'
   import { useI18n } from 'vue-i18n'
+  import type { LoginParams } from '@/api/_auth/model'
+  import config from '@/config'
 
   const { t } = useI18n()
   const router = useRouter()
+  const showVerifyDialog = ref(false)
   const { $notify, $message } = useMessage()
   const loginForm = reactive<LoginParams>({
     username: '',
@@ -25,24 +25,30 @@
 
   async function doLogin(formEl: FormInstance | undefined) {
     if (!formEl) return
-    await formEl.validate(async valid => {
+    await formEl.validate(async (valid) => {
       if (!valid) return false
       loading.value = true
       const [success] = await useLoginByPassword(loginForm)
       loading.value = false
       if (success) {
-        $message.success({
-          message: '登录成功'
-        })
-        router.push('/home')
-        $notify({
-          title: 'Welcome',
-          message: t('home.welcome'),
-          type: 'success',
-        })
-      } else {
+        showVerifyDialog.value = true
+      }
+      else {
         $message.error('账户名不存在或密码错误！')
       }
+    })
+  }
+
+  function handleVerifyChange(success: boolean) {
+    if (!success) return
+    $message.success({
+      message: '登录成功'
+    })
+    router.push('/home')
+    $notify({
+      title: 'Welcome',
+      message: t('home.welcome'),
+      type: 'success'
     })
   }
 </script>
@@ -64,9 +70,9 @@
       <el-form ref="loginFormRef" :model="loginForm" :rules="loginRules">
         <el-form-item prop="username">
           <el-input
-            size="large"
             v-model="loginForm.username"
-            :placeholder="$t('home.account') + ' admin'"
+            size="large"
+            :placeholder="`${$t('home.account')} admin`"
           />
         </el-form-item>
         <el-form-item prop="password" :style="{ marginTop: '38px' }">
@@ -74,14 +80,16 @@
             v-model="loginForm.password"
             type="password"
             size="large"
-            :placeholder="$t('home.password') + ' 123456'"
+            :placeholder="`${$t('home.password')} 123456`"
             show-password
             @keyup.enter="doLogin(loginFormRef)"
           />
         </el-form-item>
         <el-form-item>
-          <div w="360px" flex justify="between" items="center">
-            <el-checkbox v-model="loginForm.rememberMe">{{ $t('home.remember') }}</el-checkbox>
+          <div w="360px" flex="center" justify="between">
+            <el-checkbox v-model="loginForm.rememberMe">
+              {{ $t('home.remember') }}
+            </el-checkbox>
             <span
               text="sm primary"
               cursor="pointer"
@@ -108,17 +116,17 @@
     <div class="divider">
       <div>{{ $t('home.moreLoginType') }}</div>
     </div>
-    <div center text="1.3rem placeholder">
-      <i-ri-github-fill class="link" />
-      <i-ri-wechat-fill class="link" />
-      <i-ri-twitter-fill class="link" />
-      <i-ri-google-fill class="link" text="1.2rem!" />
+    <div flex="center" text="1.3rem placeholder">
+      <div i-ri-github-fill class="link" />
+      <div i-ri-wechat-fill class="link" />
+      <div i-ri-twitter-fill class="link" />
+      <div i-ri-google-fill class="link" text="1.2rem!" />
     </div>
+    <VerifyDialog v-model="showVerifyDialog" @change="handleVerifyChange" />
   </div>
 </template>
 
 <style lang="scss" scoped>
-
 :deep(.el-form) {
   .el-input__wrapper, .el-checkbox__inner {
     @apply bg-transparent;
@@ -134,7 +142,7 @@
 }
 
 .divider {
-  @apply flex center my-4;
+  @apply flex-center my-4;
 
   div {
     @apply px-3 text-xs text-secondary tracking-1px;